Crypto: Examples: Fixed warnings missed out due to -Werror flag not propagated in example cmake (#833)
* Crypto: Examples: Fixed warnings missed out due to -Werror flag not used in example cmake

-- Also fix all warnings from example sources


* Crypto: Examples: Clean up unused function alloc_and_test()


* Crypto: Examples cleanup: Removed unused functions from gcm examples


* Crypto: Cmake: Examples compilation: Silence false warnings (knwon issue) with GCC14


* Crypto: Examples: Silenced format-overflow warnings for gcc


* Examples:AES:CCM: Refactor CCM example

Reduce chance of memory leaks.


* Examples:AES:CFB: Refactor CFB example

Reduce chance of memory leaks.


* Examples:AES:SIV: Refactor SIV example

Reduce chance of memory leaks.


* Examples:AES:XTS: Refactor XTS example

Reduce chance of memory leaks.


* Examples:Cipher:AES-CCM: Remove duplicates added while merging


* Examples:Cipher:AES:CCM:SIV:CFB:XTS: Rename variables


---------

+++ examples/misc/filecrypt.cc
@@ -76,7 +76,7 @@ namespace utilities {
             }
             out.push_back(0x0f); // Pushing Excape, mark padding
             rem--;
-            for (int i = 0; i < rem; i++) {
+            for (Uint64 i = 0; i < rem; i++) {
                 out.push_back(0x00);
             }
             return out;
